Adicionado o servico de produto e incluido o arquivo da Collection do Postman
